refactor(rendering)Moved renderWorldModel to separate files
